+++ src/tiled/editableimagelayer.h
@@ -33,6 +33,7 @@ class EditableImageLayer : public EditableLayer
 
     Q_PROPERTY(QColor transparentColor READ transparentColor WRITE setTransparentColor)
     Q_PROPERTY(QUrl imageSource READ imageSource WRITE setImageSource)
+    Q_PROPERTY(QString imageFileName READ imageFileName WRITE setImageFileName)
     Q_PROPERTY(Tiled::ScriptImage *image READ image WRITE setImage)
     Q_PROPERTY(bool repeatX READ repeatX WRITE setRepeatX)
     Q_PROPERTY(bool repeatY READ repeatY WRITE setRepeatY)
@@ -47,12 +48,14 @@ public:
 
     const QColor &transparentColor() const;
     const QUrl &imageSource() const;
+    QString imageFileName() const;
     ScriptImage *image() const;
     bool repeatX() const;
     bool repeatY() const;
 
     void setTransparentColor(const QColor &transparentColor);
     void setImageSource(const QUrl &imageSource);
+    void setImageFileName(const QString &fileName);
     void setRepeatX(bool repeatX);
     void setRepeatY(bool repeatY);
 
@@ -70,6 +73,11 @@ inline const QColor &EditableImageLayer::transparentColor() const
 inline const QUrl &EditableImageLayer::imageSource() const
 {
     return imageLayer()->imageSource();
+}
+
+inline QString EditableImageLayer::imageFileName() const
+{
+    return imageLayer()->imageSource().toString(QUrl::PreferLocalFile);
 }
 
 inline bool EditableImageLayer::repeatX() const